feat(intro_doc): complete introduction summary
- Added a lyx file detailing each section of the introduction lesson, as
it is currently layed out in plan and presnetation.

- For the documents above, I also added two images into the Visuals
directory, detailing the speed vector of a ball as it is
thrown ("ball_throw.pdf"), and the PID control loop
diagram (pid_loop.png). Both of these are used in the second section of
the document, detailing the connection between math and robotics through
physics and control theory.

- Note that farther commits on this branch are possible, but will
probably be limited to spellchecking.